+++ R/UniprocessFuture-class.R
@@ -13,12 +13,13 @@
 UniprocessFuture <- function(expr = NULL, substitute = TRUE, envir = parent.frame(), globals = TRUE, packages = NULL, lazy = FALSE, local = TRUE, ...) {
   if (substitute) expr <- substitute(expr)
 
+  if (!isTRUE(local)) {
+    .Defunct(msg = "Using 'local = FALSE' with uniprocess future is defunct. It was deprecated in future 1.20.0 (2020-10-30).", package = .packageName)
+  }
+
+
   ## WORKAROUND: Skip scanning of globals if already done /HB 2021-01-18
   if (!isTRUE(attr(globals, "already-done", exact = TRUE))) {
-    if (lazy && !local && (!is.logical(globals) || globals)) {
-      stop("Non-supported use of lazy uniprocess futures: Whenever argument 'local' is FALSE, then argument 'globals' must also be FALSE. Lazy uniprocess future evaluation in the calling environment (local = FALSE) can only be done if global objects are resolved at the same time.")
-    }
-  
     ## Global objects?
     gp <- getGlobalsAndPackages(expr, envir = envir, tweak = tweakExpression, globals = globals)
     globals <- gp$globals
@@ -32,7 +33,7 @@ UniprocessFuture <- function(expr = NULL, substitute = TRUE, envir = parent.fram
     gp <- NULL
   }
  
-  future <- Future(expr = expr, substitute = FALSE, envir = envir, lazy = lazy, asynchronous = FALSE, local = local, globals = globals, packages = packages, ...)
+  future <- Future(expr = expr, substitute = FALSE, envir = envir, lazy = lazy, asynchronous = FALSE, local = TRUE, globals = globals, packages = packages, ...)
   future <- structure(future, class = c("UniprocessFuture", class(future)))
   future
 }
@@ -54,7 +55,10 @@ run.UniprocessFuture <- function(future, ...) {
 
   expr <- getExpression(future)
   envir <- future$envir
-  if (future$local) envir <- new.env(parent = envir)
+  if (!isTRUE(future$local)) {
+    .Defunct(msg = sprintf("Using 'local = FALSE' with %s futures is defunct", class(future)[1]))
+  }
+  envir <- new.env(parent = envir)
 
   ## Assign globals to separate "globals" enclosure environment?
   globals <- future$globals
@@ -72,7 +76,7 @@ run.UniprocessFuture <- function(future, ...) {
   ## WORKAROUND: Ditto warning by Future() is muffled for UniprocessFuture.
   ## /HB 2022-04-27
   if (!future$local) {
-    .Deprecated(msg = "Using 'local = FALSE' for a future is deprecated in future (>= 1.20.0) and will soon be defunct and produce an error.", package = .packageName)
+    .Defunct(msg = "Using 'local = FALSE' for a future is defunct. It was deprecated in future 1.20.0 (2020-10-30).", package = .packageName)
   }
 
   ## Always signal immediateCondition:s and as soon as possible.